bug: 分级管理员超时越界问题 #3673

+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/ManagerUserService.kt
@@ -111,7 +111,7 @@ class ManagerUserService @Autowired constructor(
             createUser = userId,
             managerId = managerUser.managerId,
             startTime = System.currentTimeMillis(),
-            timeoutTime = System.currentTimeMillis() + (DateTimeUtil.minuteToSecond(managerUser.timeout!!) * 1000),
+            timeoutTime = System.currentTimeMillis() + (DateTimeUtil.minuteToSecond(managerUser.timeout!!).toLong() * 1000),
             userId = managerUser.userId
         )
